Move compiler code around.
Make room for another compiler implementation, and add an interface.
