+++ src/AssetCompiler.php
@@ -11,65 +11,5 @@
  * @since         0.0.1
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace MiniAsset;
-
-use MiniAsset\Filter\FilterRegistry;
-use Cake\Core\Configure;
-use RuntimeException;
-
-/**
- * Compiles a set of assets together, and applies filters.
- * Forms the center of MiniAsset
- */
-class AssetCompiler
-{
-    /**
-     * The filter registry to use.
-     *
-     * @var \MiniAsset\FilterRegistry
-     */
-    protected $filterRegistry;
-
-    /**
-     * Set to true when in development mode.
-     *
-     * Enabling this disables output filters.
-     *
-     * @var bool
-     */
-    protected $debug = false;
-
-    /**
-     * Constructor.
-     *
-     * @param FilterRegistry $filters The filter registry
-     * @return void
-     */
-    public function __construct(FilterRegistry $filters, $debug)
-    {
-        $this->filterRegistry = $filters;
-        $this->debug = $debug;
-    }
-
-    /**
-     * Generate a compiled asset, with all the configured filters applied.
-     *
-     * @param AssetTarget $target The target to build
-     * @return The processed result of $target and it dependencies.
-     * @throws RuntimeException
-     */
-    public function generate(AssetTarget $build)
-    {
-        $filters = $this->filterRegistry->collection($build);
-        $output = '';
-        foreach ($build->files() as $file) {
-            $content = $file->contents();
-            $content = $filters->input($file->path(), $content);
-            $output .= $content . "\n";
-        }
-        if (!$this->debug || php_sapi_name() === 'cli') {
-            $output = $filters->output($build->path(), $output);
-        }
-        return trim($output);
-    }
-}
+// @deprecated Backwards compatibility alias
+class_alias('MiniAsset\Output\Compiler', 'MiniAsset\AssetCompiler');
